pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "kln-ai"
version = "1.0.0b4"
description = "Multi-model code review and knowledge capture system for Claude Code"
readme = "README.md"
license = {text = "Apache-2.0"}
requires-python = ">=3.9"
authors = [
    {name = "Calin Faja", email = "calinfaja@gmail.com"}
]
keywords = ["claude", "code-review", "ai", "llm", "cli", "knowledge-base"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: POSIX :: Linux",
    "Operating System :: MacOS",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Quality Assurance",
    "Topic :: Software Development :: Testing",
]
dependencies = [
    # Core CLI
    "click>=8.0.0",
    "rich>=13.0.0",
    "pyyaml>=6.0.0",
    "httpx>=0.27.0",
    # Knowledge DB - lightweight fastembed (ONNX-based, ~200MB vs 7GB with txtai)
    "fastembed>=0.3.0",
    "numpy>=1.24.0",
    # SmolKLN Agents (required for agents and kln multi)
    "smolagents[litellm]>=1.17.0",
    # LiteLLM proxy dependencies (backoff, uvicorn, etc.) for kln start
    "litellm[proxy]>=1.0.0",
    "ddgs>=6.0.0",
    "markdownify>=0.11.0",
    "lizard>=1.17.0",
]

[project.optional-dependencies]
# Optional: Claude Agent SDK for building custom agents
agent-sdk = [
    "anthropic>=0.34.0,<1.0.0",
]
# Optional: Phoenix telemetry for agent tracing
telemetry = [
    "arize-phoenix>=4.0",
    "opentelemetry-sdk",
    "opentelemetry-exporter-otlp",
    "openinference-instrumentation-smolagents",
]
# Optional: Toon integration
toon = [
    "python-toon>=0.1.0",
]
# All optional extras
all = [
    "kln-ai[agent-sdk,telemetry,toon]",
]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
kln = "klean.cli:main"
kln-smol = "klean.smol.cli:main"

[project.urls]
Homepage = "https://github.com/calinfaja/k-lean"
Documentation = "https://github.com/calinfaja/k-lean#readme"
Repository = "https://github.com/calinfaja/k-lean"
Issues = "https://github.com/calinfaja/k-lean/issues"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
klean = [
    "data/**/*",
]

# Ruff configuration
[tool.ruff]
line-length = 100
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "C4"]
ignore = ["E501"]

# Black configuration
[tool.black]
line-length = 100
target-version = ["py39", "py310", "py311", "py312"]

# Pytest configuration
[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
